% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurofeedback.R
\name{print.NIRSNeurofeedback}
\alias{print.NIRSNeurofeedback}
\title{Display a governed NIRS neurofeedback runtime}
\usage{
\method{print}{NIRSNeurofeedback}(x, ...)
}
\arguments{
\item{x}{A \code{NIRSNeurofeedback} runtime.}

\item{...}{Unused.}
}
\description{
Display a governed NIRS neurofeedback runtime
}
